Reorganize directory structures, modularize server
