+++ 06-Eventonica-SQL/.gitignore
@@ -1,4 +1,4 @@
 keys.js
 node_modules
-server
 .vscode
+db.sql
