+++ 06-Eventonica-SQL/error.html
@@ -18,7 +18,6 @@
 
     <!-- JS Files -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
-    <script src="eventRecommender.js"></script>
     <script src="index.js"></script>
     <script src="jQuery-scripts.js"></script>
