+++ 06-Eventonica-SQL/helpers/functions.js
