+++ 06-Eventonica-SQL/public/index.html
@@ -19,7 +19,6 @@
     <!-- JS Files -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
     <script src="keys.js"></script>
-    <script src="eventRecommender.js"></script>
     <script src="jQuery-scripts.js"></script>
 
     <title>Hello, world!</title>
