+++ 06-Eventonica-SQL/server/api.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+
+router.use("/events", require("./routes/events"));
+router.use("/users", require("./routes/users"));
+
+module.exports = router;
