--- 06-Eventonica-SQL/utils/data.json
@@ -1,1 +1,0 @@
-{"events":[{"name":"Drag Diva Brunch: Madonna","date":"2020-03-14T18:00:00.000Z","category":"Arts & Theatre","id":"G5vYZp7D5CI0E","image":"https://s1.ticketm.net/dbimages/375986a.jpg","description":"17 & under must be accompanied by a parent or guardian. 18 & over must have valid, government-issued photo ID. All non-headliner lineups are subject to change. General Admission includes 1 entrée and first free drink (Bloody Mary, or Mimosa for 21+ and juice, coffee, or tea for under 21): $39.00","url":"https://concerts.livenation.com/drag-diva-brunch-madonna-sacramento-california-03-14-2020/event/1C00584ECB3A8927","city":"Sacramento, CA","venue":"2100 Arden Way","dateAdded":1582763298255},{"name":"Glass Animals","date":"2020-03-12T02:00:00.000Z","category":"Music","id":"Z7r9jZ1Ae0Y8N","image":"https://s1.ticketm.net/dam/a/137/0b676399-3136-469b-add1-33582f0b1137_797691_RETINA_PORTRAIT_16_9.jpg","url":"http://www.ticketsnow.com/InventoryBrowse/TicketList.aspx?PID=2930423","city":"West Hollywood, CA","venue":"9081 Santa Monica Blvd.","dateAdded":1582763328647},{"name":"Winter Hackathon","date":"2020-02-21T00:00:00.000Z","category":"Indeed Events","id":"123","dateAdded":1582763377656},{"name":"Glass Animals","date":"2020-03-20T00:00:00.000Z","category":"Music","id":"G5dIZp7hE7IW6","image":"https://s1.ticketm.net/dam/a/137/0b676399-3136-469b-add1-33582f0b1137_797691_RETINA_PORTRAIT_16_9.jpg","description":"Doors open at 7:00pm. Ticket text reflects door time. This show is all ages. Everyone requires a ticket regardless of age. All support acts subject to change without notice. Please note that the Mezzanine and Balcony are only accessible by stairs. Patrons needing accessible seating should call the box office.","url":"https://concerts.livenation.com/glass-animals-san-antonio-texas-03-19-2020/event/3A005846E7128F47","city":"San Antonio, TX","venue":"104 North St. Mary's Street","dateAdded":1582769263523},{"name":"Glass Animals","date":"2020-06-10T23:30:00.000Z","category":"Music","id":"vv1AAZA-6GkeUOkyk","image":"https://s1.ticketm.net/dam/a/137/0b676399-3136-469b-add1-33582f0b1137_797691_RETINA_PORTRAIT_16_9.jpg","description":"Egyptian Room at Old National Centre All ages. Doors: 6:30 pm Show: 7:30 pm Each online ticket purchased includes one (1) physical CD copy of a forthcoming Glass Animals music release. After your ticket purchase you'll receive an additional email in the coming weeks with instructions on how to redeem your album. US/CAN residents only. One CD per online ticket. Not valid for Fan to Fan Resale.","url":"https://concerts.livenation.com/glass-animals-indianapolis-indiana-06-10-2020/event/05005847B6806E46","city":"Indianapolis, IN","venue":"502 N. New Jersey","dateAdded":1582769265188},{"name":"Bonnaroo Music + Arts Festival","date":null,"category":"Music","id":"G5viZ4M2ackwK","image":"https://s1.ticketm.net/dam/a/433/6c0fcbc5-6ecf-43e5-8b8d-5766494af433_1246051_RETINA_PORTRAIT_16_9.jpg","url":"https://on.fgtix.com/trk/Mq4t","city":"Manchester, TN","venue":"Between Exits 111 & Exit 114 on I-24","dateAdded":1582769278501},{"name":{"name":"Event 1","id":"1","dateAdded":1582931250049},"date":null,"id":71834,"dateAdded":1582931250049},{"name":{"name":"Event 2","id":"2","dateAdded":1582931261725},"date":null,"id":28903,"dateAdded":1582931261725},{"name":{"name":"Event 2","id":"2","dateAdded":1582931280428},"date":null,"id":10413,"dateAdded":1582931280428}],"users":[{"name":"2","personalEvents":[],"id":"1"},{"name":"2","personalEvents":[],"id":"12"},{"name":"2","personalEvents":[],"id":"3"}]}
